Sugerencias implementadas
