--- src/main/java/org/bootcamp/javazoo/service/interfaces/ICountService.java
@@ -1,8 +1,0 @@
-package org.bootcamp.javazoo.service.interfaces;
-
-import org.bootcamp.javazoo.dto.response.CountFollowersDto;
-import org.bootcamp.javazoo.dto.response.FollowersListDto;
-
-public interface ICountService {
-    CountFollowersDto getFollowersListCount(Integer userId);
-}
